refactor(portal-framework-ui): extract AllowStepNavigation type and fix import paths
- Create navigation.ts with AllowStepNavigation type definition
- Export navigation types from shared/types/index.ts
- Replace inline `boolean | (() => boolean)` with AllowStepNavigation in form/types.ts
- Replace inline `boolean | (() => boolean)` with AllowStepNavigation in shared/types/header.ts
- Fix incorrect relative import paths in footer.ts and header.ts
