+++ libs/portal-framework-ui/src/components/shared/types/index.ts
@@ -3,4 +3,5 @@ export * from "./environment";
 export * from "./footer";
 export * from "./form";
 export * from "./header";
+export * from "./navigation";
 export * from "./step";
