+++ libs/portal-framework-ui/src/components/shared/types/navigation.ts
@@ -1,0 +1,7 @@
+import type { BaseRecord } from "@refinedev/core";
+
+/**
+ * Type for controlling step navigation in wizards
+ * Can be a boolean value or a function that returns a boolean
+ */
+export type AllowStepNavigation = boolean | (() => boolean);
